Abstracted dendrite functions
The new file, dendrite_functions.py, will hold all created functions and permutations of said functions for purposes of testing the capabilities of dendrites.  At the bottom of the file is a "configuration" section, where new dendrite functions should be stored.  When a different dendrite function is desired, simply change the function number in model.py.